resolve build warning

+++ src/conn/conn_log.c
@@ -311,8 +311,8 @@ __wt_logmgr_config(WT_SESSION_IMPL *session, const char **cfg, bool reconfig)
     WT_RET(__wt_config_gets(session, cfg, "log.prealloc", &cval));
     if (cval.val != 0) {
         WT_RET(__wt_config_gets(session, cfg, "log.prealloc_init_count", &cval));
-        conn->log_prealloc = cval.val;
-        conn->log_prealloc_init_count = cval.val;
+        conn->log_prealloc = (uint32_t)cval.val;
+        conn->log_prealloc_init_count = (uint32_t)cval.val;
         WT_ASSERT(session, conn->log_prealloc > 0);
     }
